Refactor ScrollDiscover to its own folder
