+++ src/components/ScrollDiscover/ScrollDiscover.tsx
@@ -1,6 +1,6 @@
-import "./scrollDiscover.css";
+import './ScrollDiscover.css';
 
 export const ScrollDiscover = () => {
-  const text = "SCROLL TO\n\n\n\n\n\nDISCOVER";
+  const text = 'SCROLL TO\n\n\n\n\n\nDISCOVER';
   return <div className="scroll-discover">{text}</div>;
 };
